frontend/requirements.txt
backports.functools-lru-cache>=1.5
backports.ssl-match-hostname>=3.5.0.1
blessed>=1.14.2
botocore>=1.10.28
cement>=2.8.2
certifi>=2018.4.16
chardet>=3.0.4
colorama>=0.3.7
configparser>=3.5.0
coreapi>=2.3.3
coreschema>=0.0.4
cycler>=0.10.0
Django == 2.0.7
django-allauth>=0.36.0
django-filter>=1.1.0
djangorestframework>=3.8.2
docker>=2.6.1
docker-pycreds>=0.2.3
dockerpty>=0.4.1
docopt>=0.6.2
docutils>=0.14
ez-setup>=0.9
idna>=2.6
ipaddress>=1.0.22
itypes>=1.1.0
Jinja2>=2.10
jmespath>=0.9.3
Markdown>=2.6.11
MarkupSafe>=1.0
matplotlib>=2.1.2
mysqlclient>=1.3.12
numpy>=1.14.2
oauthlib>=2.1.0
pathspec>=0.5.5
Pillow>=5.1.0
pyparsing>=2.2.0
python-dateutil>=2.6.1
python-openid>=2.2.5
pytz>=2018.3
PyYAML>=3.12
requests>=2.9.1
requests-oauthlib>=0.8.0
scipy>=1.0.0
semantic-version>=2.5.0
six>=1.11.0
subprocess32>=3.2.7
tabulate>=0.7.5
termcolor>=1.1.0
uritemplate>=3.0.0
urllib3>=1.22
virtualenv>=16.0.0
wcwidth>=0.1.7
websocket-client>=0.48.0
pytest
pytest-django
pytest-cov
mixer
ipdb
